Small fixes

+++ src/ku/SvgUtilities.cpp
@@ -288,7 +288,7 @@ std::string SvgUtilities::getPointString(std::pair<double, double> point) {
 std::string SvgUtilities::padLeft(const std::string& input, char character, int32_t n) {
   // return std::string(n - std::min(static_cast<size_t>(n), input.length()), character) + input;
 
-  if (n > input.size()) {
+  if (static_cast<size_t>(n) > input.size()) {
     std::string output = input;
     output.insert(0, n - input.size(), character);
     return output;
